QueryDSL AOP 적용, WebConfig 추가, 요청 URL 추가(세종 특별시)

+++ src/main/java/kr/ac/skuniv/realestate/domain/entity/RentDate.java
@@ -20,10 +20,10 @@ public class RentDate {
     private Date date;
 
     @Column(name = "guarantee_price")
-    private Long guaranteePrice;
+    private Double guaranteePrice;
 
     @Column(name = "monthly_price")
-    private Long monthlyPrice;
+    private Double monthlyPrice;
 
     @ManyToOne(cascade = {CascadeType.ALL}, fetch = FetchType.LAZY)
     @JoinColumn(name = "buildingNo")
